feat: location handler added | real name handler added | time keyboard added

+++ src/engine/telegram/models.py
@@ -35,7 +35,7 @@ class PaySystem(models.Model):
 
 class User(models.Model):
 
-    chat_id = models.IntegerField(
+    chat_id = models.BigIntegerField(
         'Chat id',
         default=0,
         blank=False,
@@ -80,7 +80,7 @@ class User(models.Model):
         null=True
         )
 
-    phone = models.PositiveIntegerField(
+    phone = models.BigIntegerField(
         'Phone',
         default=None,
         blank=True,
